Improve user-friendliness and readability

+++ order.sh
@@ -1,9 +1,11 @@
 STREAM="$1"
 
-mkdir ordered || rm ordered/* || :
+if [ -z "$NO_RM" ]; then
+  mkdir ordered || rm ordered/* || :
+fi
 
 if [ -z "$STREAM" ]; then
-  files="$(find "képek/" -name "stats.txt" -newer "képek/2021-07-15_GlueRab7_ctrl_-2h/GlueRab7_ctrl_-2h-0018.tif_Files")"
+  files="$(find "képek/" -name "stats.txt" -newer "képek/2021-07-15_GlueRab7_ctrl_-2h/GlueRab7_ctrl_-2h-0018.tif_Files/stats.txt")"
   count="$(echo "$files" | wc -l)"
 else
   count="?"
@@ -17,19 +19,26 @@ else
   cat "$STREAM"
 fi | while IFS= read file; do
   let i++
-  printf "$i / $count\r"
   if [ -z "$STREAM" ]; then
     folder="../$(dirname "$file")"
   else
     folder="$file"
     file="$folder/stats.txt"
   fi
+  
+  if [ -n "$NO_RM" ]; then
+    if [ -n "$(find ordered/ -name "*-$(basename "$folder")" -maxdepth 1 -print -quit)" ]; then
+      continue
+    fi
+  fi
+
+  printf "$i / $count\r"
 
   if [ $(cat "$file" | wc -l) -lt 2 ]; then
     echo "Warning: $file is invalid - skipping"
     continue
   fi
-  linkname="ordered/$(cat "$file" | head -n1)-$(basename "$folder")"
+  linkname="ordered/$(cat "$file" | grep -oP 'Scalar positives: \K.*')-$(basename "$folder")"
    ln -s "$folder/composite.tif" "$linkname.tif"
    ln -s "$folder" "$linkname"
   # convert "$folder"/*_c?.tif -combine "$linkname"
